Renamed test-data folder and imports to testData. Deleted all .npmignore files. Added 'files' field to package.json files, filled the array with whats needed to publish.
