+++ packages/create-sitecore-jss/.eslintrc
@@ -1,10 +1,4 @@
 {
-  "extends": [
-    "../../.eslintrc",
-    "../../eslint-configs/typescript"
-  ],
-  "ignorePatterns": [
-    "src/templates/**/*", 
-    "src/common/test-data/**/*"
-  ]
+  "extends": ["../../.eslintrc", "../../eslint-configs/typescript"],
+  "ignorePatterns": ["src/templates/**/*", "src/common/testData/**/*"]
 }
