+++ packages/create-sitecore-jss/.nycrc
@@ -6,7 +6,7 @@
     "**/*.d.ts",
     "**/*.test.ts",
     "src/templates/**/*",
-    "src/common/test-data",
+    "src/common/testData",
     "dist"
   ],
   "all": true,
@@ -14,5 +14,7 @@
     "json-summary",
     "text"
   ],
-  "require": ["ts-node/register"]
+  "require": [
+    "ts-node/register"
+  ]
 }
