+++ packages/create-sitecore-jss/src/common/testData/pkg.ts
