+++ packages/create-sitecore-jss/src/common/testData/test.package.json
