+++ packages/create-sitecore-jss/src/common/testData/transform/source.json
