+++ packages/create-sitecore-jss/src/common/testData/transform/target.ts
