+++ packages/create-sitecore-jss/tsconfig.json
@@ -16,6 +16,6 @@
     "src/templates/**/*",
     "src/test",
     "src/**/*.test.ts",
-    "src/**/test-data/**/*"
+    "src/**/testData/**/*"
   ]
 }
