+++ packages/sitecore-jss-angular/.npmignore
@@ -1,18 +1,2 @@
-src
-testData
-coverage
-sitecore
-index.test.js
-.eslintrc
-.vscode
-.npmignore
-.gitignore
-jsconfig.json
-tsconfig.json
-ng-package.json
-karma.conf.js
-npm-debug*
 /dist/package.json
 /dist/README.md
-*.tgz
-.nycrc
