--- packages/sitecore-jss-cli/.npmignore
@@ -1,15 +1,0 @@
-src
-testData
-sitecore
-index.test.js
-.vscode
-.babelrc
-.eslintrc
-.npmignore
-.gitignore
-jsconfig.json
-tsconfig.json
-npm-debug*
-!/types
-.nycrc
-*.tgz
