+++ packages/sitecore-jss-cli/package.json
@@ -63,5 +63,9 @@
     "ts-node": "^8.4.1",
     "typescript": "~4.3.5"
   },
-  "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b"
+  "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b",
+  "files": [
+    "dist",
+    "types"
+  ]
 }
