+++ packages/sitecore-jss-forms/package.json
@@ -48,5 +48,9 @@
   },
   "description": "",
   "types": "types/index.d.ts",
-  "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b"
+  "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b",
+  "files": [
+    "dist",
+    "types"
+  ]
 }
