--- packages/sitecore-jss-nextjs/.npmignore
@@ -1,19 +1,0 @@
-src
-testData
-sitecore
-index.test.js
-.vscode
-.babelrc
-.eslintrc
-.npmignore
-.gitignore
-jsconfig.json
-tsconfig.json
-tsconfig-esm.json
-npm-debug*
-*.log
-!/types
-.nycrc
-.nyc_output
-coverage
-*.tgz
