+++ packages/sitecore-jss-react-forms/package.json
@@ -60,5 +60,9 @@
   },
   "description": "",
   "types": "types/index.d.ts",
-  "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b"
+  "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b",
+  "files": [
+    "dist",
+    "types"
+  ]
 }
