+++ packages/sitecore-jss-react-native/package.json
@@ -69,5 +69,8 @@
     "babel-core": "6.26.3"
   },
   "types": "./types/index.d.ts",
-  "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b"
+  "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b",
+  "files": [
+    "dist"
+  ]
 }
