--- packages/sitecore-jss-react/.npmignore
@@ -1,16 +1,0 @@
-src
-testData
-sitecore
-index.test.js
-.vscode
-.babelrc
-.eslintrc
-.npmignore
-.gitignore
-jsconfig.json
-tsconfig.json
-npm-debug*
-*.log
-!/types
-.nycrc
-*.tgz
