+++ packages/sitecore-jss-rendering-host/package.json
@@ -61,5 +61,9 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Sitecore/jss.git"
-  }
+  },
+  "files": [
+    "dist",
+    "types"
+  ]
 }
