--- packages/sitecore-jss-vue/.npmignore
@@ -1,19 +1,0 @@
-src
-testData
-sitecore
-*.test.js
-.vscode
-.babelrc
-.eslintrc
-.npmignore
-.gitignore
-jsconfig.json
-tsconfig.json
-npm-debug*
-*.log
-jest.config.js
-__snapshots__
-.prettierignore
-!/types
-.nycrc
-*.tgz
