+++ packages/sitecore-jss-vue/package.json
@@ -53,5 +53,9 @@
     "@vue/compiler-sfc": "^3.0.11"
   },
   "types": "./types/index.d.ts",
-  "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b"
+  "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b",
+  "files": [
+    "dist",
+    "types"
+  ]
 }
