+++ packages/sitecore-jss/package.json
@@ -64,5 +64,11 @@
   },
   "description": "",
   "types": "types/index.d.ts",
-  "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b"
+  "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b",
+  "files": [
+    "dist",
+    "types",
+    "./*.js",
+    "./*.d.ts"
+  ]
 }
