+++ packages/sitecore-jss/src/graphql/graphql.test.ts
@@ -2,7 +2,7 @@
 import { expect } from 'chai';
 import { getAppRootId, siteNameError, languageError } from './app-root-query';
 import { GraphQLRequestClient } from './../graphql-request-client';
-import appRootQueryResponse from '../test-data/mockAppRootQueryResponse.json';
+import appRootQueryResponse from '../testData/mockAppRootQueryResponse.json';
 import nock from 'nock';
 import { SitecoreTemplateId } from '../constants';
