+++ packages/sitecore-jss/src/testData/mockAppRootQueryResponse.json
